♻️  peer: Split into modules
This creates a new hierarchy for peer mod. We'll be adding more
freedesktop interfaces to the peer soon, so it would be good to have
everything separated out.
