+++ src/peer/mod.rs
@@ -1,21 +1,20 @@
-use std::collections::HashSet;
+mod fdo;
 
 use anyhow::Result;
-use enumflags2::BitFlags;
 use tracing::trace;
 use zbus::{
-    dbus_interface,
-    fdo::{self, ReleaseNameReply, RequestNameFlags, RequestNameReply},
-    names::{BusName, OwnedBusName, OwnedUniqueName, OwnedWellKnownName},
-    AuthMechanism, Connection, ConnectionBuilder, Guid, MessageStream, OwnedMatchRule, Socket,
+    names::{BusName, OwnedUniqueName},
+    AuthMechanism, Connection, ConnectionBuilder, Guid, MessageStream, Socket,
 };
 
 use crate::name_registry::NameRegistry;
+use fdo::DBus;
 
 /// A peer connection.
 #[derive(Debug)]
 pub struct Peer {
     conn: Connection,
+    name_registry: NameRegistry,
     unique_name: OwnedUniqueName,
 }
 
@@ -34,7 +33,7 @@ impl Peer {
             .p2p()
             .serve_at(
                 "/org/freedesktop/DBus",
-                DBus::new(unique_name.clone(), name_registry),
+                DBus::new(unique_name.clone(), name_registry.clone()),
             )?
             .name("org.freedesktop.DBus")?
             .unique_name("org.freedesktop.DBus")?
@@ -43,7 +42,11 @@ impl Peer {
             .await?;
         trace!("created: {:?}", conn);
 
-        Ok(Self { conn, unique_name })
+        Ok(Self {
+            conn,
+            name_registry,
+            unique_name,
+        })
     }
 
     pub fn unique_name(&self) -> &OwnedUniqueName {
@@ -71,7 +74,7 @@ impl Peer {
         let dbus = dbus_ref.get().await;
         let hdr = msg.header().expect("received message without header");
 
-        dbus.match_rules.iter().any(|rule| {
+        let ret = dbus.match_rules().any(|rule| {
             // First make use of zbus API
             match rule.matches(msg) {
                 Ok(false) => return false,
@@ -85,7 +88,7 @@ impl Peer {
 
             // Then match sender and destination involving well-known names, manually.
             if let Some(sender) = rule.sender().cloned().and_then(|name| match name {
-                BusName::WellKnown(name) => dbus.name_registry.lookup(name).as_deref().cloned(),
+                BusName::WellKnown(name) => self.name_registry.lookup(name).as_deref().cloned(),
                 // Unique name is already taken care of by the zbus API.
                 BusName::Unique(_) => None,
             }) {
@@ -108,7 +111,7 @@ impl Peer {
                     .expect("DESTINATION field unset")
                     .clone()
                 {
-                    BusName::WellKnown(name) => match dbus.name_registry.lookup(name) {
+                    BusName::WellKnown(name) => match self.name_registry.lookup(name) {
                         Some(name) if name == *destination => (),
                         Some(_) => return false,
                         None => return false,
@@ -119,83 +122,8 @@ impl Peer {
             }
 
             true
-        })
-    }
-}
+        });
 
-#[derive(Debug)]
-struct DBus {
-    greeted: bool,
-    unique_name: OwnedUniqueName,
-    name_registry: NameRegistry,
-    match_rules: HashSet<OwnedMatchRule>,
-}
-
-impl DBus {
-    fn new(unique_name: OwnedUniqueName, name_registry: NameRegistry) -> Self {
-        Self {
-            greeted: false,
-            unique_name,
-            name_registry,
-            match_rules: HashSet::new(),
-        }
-    }
-}
-
-#[dbus_interface(interface = "org.freedesktop.DBus")]
-impl DBus {
-    /// Returns the unique name assigned to the connection.
-    async fn hello(&mut self) -> fdo::Result<OwnedUniqueName> {
-        if self.greeted {
-            return Err(fdo::Error::Failed(
-                "Can only call `Hello` method once".to_string(),
-            ));
-        }
-        self.greeted = true;
-
-        Ok(self.unique_name.clone())
-    }
-
-    /// Ask the message bus to assign the given name to the method caller.
-    fn request_name(
-        &self,
-        name: OwnedWellKnownName,
-        flags: BitFlags<RequestNameFlags>,
-    ) -> RequestNameReply {
-        self.name_registry
-            .request_name(name, self.unique_name.clone(), flags)
-    }
-
-    /// Ask the message bus to release the method caller's claim to the given name.
-    fn release_name(&self, name: OwnedWellKnownName) -> ReleaseNameReply {
-        self.name_registry
-            .release_name(name.into(), (&*self.unique_name).into())
-    }
-
-    /// Returns the unique connection name of the primary owner of the name given.
-    fn get_name_owner(&self, name: OwnedBusName) -> fdo::Result<OwnedUniqueName> {
-        match name.into_inner() {
-            BusName::WellKnown(name) => self.name_registry.lookup(name).ok_or_else(|| {
-                fdo::Error::NameHasNoOwner("Name is not owned by anyone. Take it!".to_string())
-            }),
-            // FIXME: Not good enough. We need to check if name is actually owned.
-            BusName::Unique(name) => Ok(name.into()),
-        }
-    }
-
-    /// Adds a match rule to match messages going through the message bus
-    fn add_match(&mut self, rule: OwnedMatchRule) {
-        self.match_rules.insert(rule);
-    }
-
-    /// Removes the first rule that matches.
-    fn remove_match(&mut self, rule: OwnedMatchRule) -> fdo::Result<()> {
-        if !self.match_rules.remove(&rule) {
-            return Err(fdo::Error::MatchRuleNotFound(
-                "No such match rule".to_string(),
-            ));
-        }
-
-        Ok(())
+        ret
     }
 }
